+++ js-packages/@prestojs/ui/src/__tests__/TimeFormatter.test.tsx
@@ -1,16 +1,42 @@
+import { render } from '@testing-library/react';
+import React from 'react';
 import TimeFormatter from '../formatters/TimeFormatter';
 
 test('TimeFormatter format times correctly', () => {
-    expect(TimeFormatter({ value: null })).toBeNull();
-    expect(TimeFormatter({ value: '1' })).toBeNull();
-    expect(TimeFormatter({ value: 'midnight' })).toBeNull();
-    // we by default does not test locales due to most Node gets built without Intl. If you wish to test locales, install full-icu package and run `NODE_ICU_DATA=node_modules/full-icu jest`.
-    expect(TimeFormatter({ value: '11:30' })).toBe('11:30:00 AM');
-    expect(TimeFormatter({ value: '22:30' })).toBe('10:30:00 PM');
-    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-    // @ts-ignore
-    expect(TimeFormatter({ value: '11:30:00', localeOptions: { timeStyle: 'short' } })).toBe(
-        '11:30 AM'
+    expect(render(<TimeFormatter value={null} />).container.innerHTML).toBe('');
+    expect(render(<TimeFormatter value={null} blankLabel="None" />).container.innerHTML).toBe(
+        'None'
     );
-    expect(TimeFormatter({ value: '11:30:00.123' })).toBe('11:30:00 AM');
+    expect(render(<TimeFormatter value="1" />).container.innerHTML).toBe('');
+    expect(
+        render(<TimeFormatter value="1" invalidValueLabel="Bad Time" />).container.innerHTML
+    ).toBe('Bad Time');
+    expect(render(<TimeFormatter value="11:30" />).container.innerHTML).toBe('11:30:00 AM');
+    expect(render(<TimeFormatter value="22:30:55" />).container.innerHTML).toBe('10:30:55 PM');
+    expect(
+        render(
+            <TimeFormatter
+                value="11:30:00.000+09:30"
+                localeOptions={{ timeStyle: 'short', timeZone: 'Australia/Adelaide' }}
+            />
+        ).container.innerHTML
+    ).toBe('11:30 AM');
+
+    expect(
+        render(
+            <TimeFormatter
+                value="11:30:00.000+09:30"
+                localeOptions={{ timeStyle: 'short', timeZone: 'Australia/Melbourne' }}
+            />
+        ).container.innerHTML
+    ).toBe('12:00 PM');
+
+    expect(
+        render(
+            <TimeFormatter
+                value={new Date(Date.parse('1970-01-01T06:16:35.720+04:00'))}
+                localeOptions={{ timeStyle: 'short', timeZone: 'Australia/Melbourne' }}
+            />
+        ).container.innerHTML
+    ).toBe('12:16 PM');
 });