+++ js-packages/@prestojs/ui/src/formatters/NumberFormatter.tsx
@@ -1,27 +1,66 @@
-/*
- * Formats a numeric input based on user browser's locale. Returns null when value is either null or NaN.
- *
- * Both `locales` and `options` used are customizeable. For values available, see: [MDN Documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/toLocaleString)
- */
-export default function NumberFormatter({
-    value,
-    locales = [],
-    localeOptions,
-}: {
-    value: number | null;
+import React, { ReactNode } from 'react';
+
+type NumberFormatterProps = {
+    /**
+     * The number to format. If `value` is a `string` then value will be coerced to `Number` and if the result is `NaN` then
+     * `invalidValueLabel` will be returned;
+     */
+    value?: number | string | null;
+    /**
+     * The [locales](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/toLocaleString#using_locales) option passed to
+     * [Number.toLocaleString](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/toLocaleString).
+     */
     locales?: string | Array<string>;
+    /**
+     * The [localeOptions](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/toLocaleString#using_options) passed to
+     * [Number.toLocaleString](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/toLocaleString).
+     */
     localeOptions?: Intl.NumberFormatOptions;
-}): string | null {
-    if (value === null) {
-        return value;
+    /**
+     * What to render when `value` is `null`, `undefined` or an empty string
+     *
+     * Defaults to `null`
+     */
+    blankLabel?: ReactNode;
+    /**
+     * What to render when passed value is not a valid number
+     *
+     * Defaults to `null`
+     */
+    invalidValueLabel?: ReactNode;
+};
+
+/**
+ * Formats a numeric input based on user browser's locale.
+ *
+ * If no value is provided `blankLabel` is returned.
+ *
+ * If an invalid value is provided `invalidValueLabel` is returned.
+ *
+ * This is the [default formatter](doc:getFormatterForField) used for [NumberField](doc:NumberField), [DecimalField](doc:DecimalField),
+ * [CurrencyField](doc:CurrencyField), [FloatField](doc:FloatField) and [IntegerField](doc:IntegerField)
+ *
+ * @extract-docs
+ * @menu-group Formatters
+ */
+export default function NumberFormatter(props: NumberFormatterProps): React.ReactElement {
+    const {
+        value,
+        locales = [],
+        blankLabel = null,
+        invalidValueLabel = null,
+        localeOptions,
+    } = props;
+    if (value == null || value === '') {
+        return <>{blankLabel}</>;
     }
     let finalValue = value;
     if (typeof value != 'number') {
         finalValue = Number(value);
     }
     if (Number.isNaN(finalValue)) {
-        return null;
+        return <>{invalidValueLabel}</>;
     }
 
-    return finalValue.toLocaleString(locales, localeOptions);
+    return <>{finalValue.toLocaleString(locales, localeOptions)}</>;
 }